ML/agents/llamaindex/requirements.txt
llama-index>=0.11.0
llama-index-llms-openai>=0.2.0
openai>=1.0.0
asyncio
python-dotenv
